feature: myinfo class 웹으로 넘기기 완료
