+++ evolve/templates/result-user-lecture.html
@@ -1,3 +1,7 @@
-{% extends 'base.html' %}
+{% extends "base.html" %}
+{% block content %}
+<h1 style="color:blue">전공 및 교양 현황</h1>
+<br>
 
-{{complete}}
+</form>
+{% endblock %}
